Refine design token overrides with palette gen and update tests
