+++ src/v3/src/util/theme.test.ts
@@ -13,59 +13,37 @@
 import * as Tokens from '@okta/odyssey-design-tokens';
 import chroma from 'chroma-js';
 
-import { createTheme, generatePalette } from './theme';
+import { createThemeAndTokens, generatePalette } from './theme';
+
+const WHITE_HEX = '#ffffff';
+// Odyssey-defined contrast ratios for their WCAG-friendly palettes
+const ODYSSEY_RATIOS = [1.1, 1.31, 1.61, 2.22, 3.32, 4.5, 4.95, 8.72, 11.73, 14.94];
 
 describe('theme utilities', () => {
   describe('generatePalette', () => {
-    it('for light primary color', () => {
-      const palette = generatePalette('#cc0000');
+    it('generates a palette with Odyssey contrast ratios', () => {
+      const palette = generatePalette('#546BE7');
       expect(palette).toBeDefined();
-      expect(palette?.main).toBe('#cc0000');
-      expect(palette?.light).toBe('#f28f78');
-      expect(palette?.lighter).toBe('#fff1f0');
-      expect(palette?.dark).toBe('#640000');
-      expect(palette?.contrastText).toBe('#ffffff');
-    });
-
-    it('for light primary color with poor contrast against white', () => {
-      const palette = generatePalette('#6666ff');
-      expect(palette).toBeDefined();
-      expect(palette?.contrastText).toBe('#1d1d21');
-    });
-
-    it('for light primary color with clamped primaryColorLight', () => {
-      const colorWithLightness60 = chroma.hsl(100, 1, 0.6);
-      const palette = generatePalette(colorWithLightness60.hex());
-
-      expect(palette).toBeDefined();
-      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-      const derivedPrimaryColorLight = chroma(palette.light!);
-      expect(derivedPrimaryColorLight.get('hsl.l')).toBeCloseTo(0.90, 2);
-    });
-
-    it('generates a derived theme for dark primary color', () => {
-      const palette = generatePalette('#170f5f');
-      expect(palette).toBeDefined();
-      expect(palette?.main).toBe('#170f5f');
-      expect(palette?.light).toBe('#bdb7f4');
-      expect(palette?.lighter).toBe('#f4f0fe');
-      expect(palette?.dark).toBe('#402ede');
-      expect(palette?.contrastText).toBe('#ffffff');
+      expect(Object.keys(palette!).length).toBe(10);
+      const contrastRatios = Object.values(palette!).map((hue) => chroma.contrast(hue, WHITE_HEX));
+      // Precision of 0.5 maps to 0.2 as an allowed deviation to account for differences between
+      // chroma.contrast() and Leonardo's calculations
+      contrastRatios.forEach((ratio, i) => expect(ratio).toBeCloseTo(ODYSSEY_RATIOS[i], 0.5));
     });
 
     it('returns null for invalid colors', () => {
       // suppress warnings from when chroma-js throws
       jest.spyOn(console, 'warn').mockImplementation(() => { });
 
-      expect(generatePalette('#12345')).toEqual({});
-      expect(generatePalette('#ff00gg')).toEqual({});
-      expect(generatePalette('')).toEqual({});
+      expect(generatePalette('#12345')).toBeNull();
+      expect(generatePalette('#ff00gg')).toBeNull();
+      expect(generatePalette('')).toBeNull();
 
       jest.restoreAllMocks();
     });
   });
 
-  describe('createTheme', () => {
+  describe('createThemeAndTokens', () => {
     // Explicitly construct OD 1.x palette because odyssey-react-mui palette's 'dark' is
     // different than Tokens.PalettePrimaryDark
     const defaultOdyPrimaryPalette = {
@@ -76,75 +54,75 @@ describe('theme utilities', () => {
       contrastText: Tokens.TypographyColorInverse,
     };
 
-    it('should use brandColors.primaryColor to set theme.palette.primary.*', () => {
+    it('should use brandColors.primaryColor to set PalettePrimary.* tokens', () => {
       const primaryColor = '#7950f2';
       const palette = generatePalette(primaryColor);
-      const theme = createTheme({ primaryColor });
-      expect(theme.palette?.primary).toEqual(palette);
+      const { tokensOverride: tokens } = createThemeAndTokens({ primaryColor });
+      // Pick a subset of the overridden primary-related design tokens to assert on
+      expect(tokens.PalettePrimaryMain).toEqual(palette?.CustomHue500);
+      expect(tokens.PalettePrimaryDark).toEqual(palette?.CustomHue700);
+      expect(tokens.TypographyColorAction).toEqual(palette?.CustomHue600);
     });
 
     it('should ignore brandColors if undefined and use default Odyssey theme', () => {
-      const theme = createTheme();
+      const { themeOverride: theme } = createThemeAndTokens();
       expect(theme.palette?.primary).toEqual(defaultOdyPrimaryPalette);
     });
 
     it('should ignore brandColors.primaryColor if invalid', () => {
-      // suppress warnings from when chroma-js throws
+      // suppress warnings from when @adobe/leonardo-contrast-color throws
       jest.spyOn(console, 'warn').mockImplementation(() => { });
       const primaryColor = '#invalid';
-      const theme = createTheme({ primaryColor });
+      const { themeOverride: theme } = createThemeAndTokens({ primaryColor });
       expect(theme.palette?.primary).toEqual(defaultOdyPrimaryPalette);
     });
 
-    it('should use PalettePrimaryMain to set theme.palette.primary.*', () => {
+    it('should use PalettePrimaryMain to set PalettePrimary.* tokens', () => {
       const PalettePrimaryMain = '#7950f2';
       const palette = generatePalette(PalettePrimaryMain);
-      const theme = createTheme(undefined, { PalettePrimaryMain });
-      expect(theme.palette?.primary).toEqual(palette);
+      const { tokensOverride: tokens } = createThemeAndTokens(undefined, { PalettePrimaryMain });
+      // Pick a subset of the overridden primary-related design tokens to assert on
+      expect(tokens.PalettePrimaryMain).toEqual(palette?.CustomHue500);
+      expect(tokens.PalettePrimaryDark).toEqual(palette?.CustomHue700);
+      expect(tokens.TypographyColorAction).toEqual(palette?.CustomHue600);
     });
 
-    it('should use PaletteDangerMain to set theme.palette.error.*', () => {
-      const PaletteDangerMain = '#e72500';
+    it('should use PaletteDangerMain to set PaletteDanger.* tokens', () => {
+      const PaletteDangerMain = '#7950f2';
       const palette = generatePalette(PaletteDangerMain);
-      const theme = createTheme(undefined, { PaletteDangerMain });
-      expect(theme.palette?.error).toEqual(palette);
+      const { tokensOverride: tokens } = createThemeAndTokens(undefined, { PaletteDangerMain });
+      // Pick a subset of the overridden danger-related design tokens to assert on
+      expect(tokens.PaletteDangerMain).toEqual(palette?.CustomHue500);
+      expect(tokens.PaletteDangerDark).toEqual(palette?.CustomHue700);
+      expect(tokens.BorderColorDangerControl).toEqual(palette?.CustomHue500);
     });
 
-    it('should use PaletteWarningMain to set theme.palette.warning.*', () => {
-      const PaletteWarningMain = '#a16c03';
+    it('should use PaletteWarningMain to set PaletteWarning.* tokens', () => {
+      const PaletteWarningMain = '#7950f2';
       const palette = generatePalette(PaletteWarningMain);
-      const theme = createTheme(undefined, { PaletteWarningMain });
-      expect(theme.palette?.warning).toEqual(palette);
+      const { tokensOverride: tokens } = createThemeAndTokens(undefined, { PaletteWarningMain });
+      // Pick a subset of the overridden warning-related design tokens to assert on
+      expect(tokens.PaletteWarningMain).toEqual(palette?.CustomHue500);
+      expect(tokens.PaletteWarningDark).toEqual(palette?.CustomHue700);
+      expect(tokens.TypographyColorWarning).toEqual(palette?.CustomHue600);
     });
 
-    it('should use PaletteSuccessMain to set theme.palette.success.*', () => {
-      const PaletteSuccessMain = '#16884a';
+    it('should use PaletteSuccessMain to set PaletteSuccess.* tokens', () => {
+      const PaletteSuccessMain = '#7950f2';
       const palette = generatePalette(PaletteSuccessMain);
-      const theme = createTheme(undefined, { PaletteSuccessMain });
-      expect(theme.palette?.success).toEqual(palette);
+      const { tokensOverride: tokens } = createThemeAndTokens(undefined, { PaletteSuccessMain });
+      // Pick a subset of the overridden success-related design tokens to assert on
+      expect(tokens.PaletteSuccessMain).toEqual(palette?.CustomHue500);
+      expect(tokens.PaletteSuccessDark).toEqual(palette?.CustomHue700);
+      expect(tokens.PaletteSuccessText).toEqual(palette?.CustomHue600);
     });
 
     it('should override brandColors.primaryColor with Tokens.PalettePrimaryMain', () => {
-      const theme = createTheme({ primaryColor: '#aaa' }, { PalettePrimaryMain: '#bbb' });
-      // @ts-expect-error Property PaletteColorOptions is a union
-      expect(theme.palette?.primary?.main).toEqual('#bbb');
-    });
-
-    it('should override generated palette colors with custom tokens', () => {
-      const customTokens = {
-        PalettePrimaryLight: '#77216F',
-        PalettePrimaryMain: '#5E2750',
-        PalettePrimaryDark: '#2C001E',
-      };
-      const palette = generatePalette('#5E2750');
-      const theme = createTheme(undefined, customTokens);
-      expect(theme.palette?.primary).toEqual({
-        lighter: palette.lighter, // generated
-        light: customTokens.PalettePrimaryLight, // provided
-        main: customTokens.PalettePrimaryMain, // provided
-        dark: customTokens.PalettePrimaryDark, // provided
-        contrastText: palette.contrastText, // generated
-      });
+      // #FF3D3D is a bright red, #7950f2 is a bright purple
+      const { tokensOverride: tokens } = createThemeAndTokens({ primaryColor: '#FF3D3D' }, { PalettePrimaryMain: '#7950f2' });
+      // PalettePrimaryMain won't be exactly #7950f2 because Leonardo does not map directly to
+      // customer's color, but #8059f3 is very similar to #7950f2
+      expect(tokens.PalettePrimaryMain).toEqual('#8059f3');
     });
 
     it('should override non-palette Odyssey default theme values with custom tokens', () => {
@@ -152,7 +130,7 @@ describe('theme utilities', () => {
         BorderRadiusMain: '1rem',
         TypographyColorHeading: '#77216F',
       };
-      const theme = createTheme(undefined, customTokens);
+      const { themeOverride: theme } = createThemeAndTokens(undefined, customTokens);
       expect(theme.mixins?.borderRadius).toEqual(customTokens.BorderRadiusMain);
       // @ts-expect-error Odyssey theme sets heading colors by default
       expect(theme.typography?.h1.color).toEqual(customTokens.TypographyColorHeading);
